Few more optimizations
Added user input of speed, abstracted some things out of main, and
reduced the iterations of the board by rolling the cell kill and gen
checks in with the function that draws the boundaries of the board.
Runs a little smoother on large boards and slow speeds are better for
small boards.